# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Archery Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: collection
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ACTIVE, ARCHIVED]
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, name_asc]
            default: newest
        - name: featured
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Product listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID or slug
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: productId
          in: query
          schema:
            type: string
        - name: approved
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Review list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    put:
      operationId: updateReview
      tags: [reviews]
      summary: Moderate a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        status:
          type: string
        shortDescription:
          type: ["string", "null"]
        price:
          type: string
        compareAtPrice:
          type: ["string", "null"]
        brandId:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, sku, status, price, isFeatured, isNewArrival, createdAt, updatedAt]
    ProductWithRelations:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        status:
          type: string
        shortDescription:
          type: ["string", "null"]
        price:
          type: string
        compareAtPrice:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        createdAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        brand:
          $ref: "#/components/schemas/BrandSummary"
        reviewCount:
          type: integer
      required: [id, name, slug, sku, status, price, isFeatured, isNewArrival, createdAt, images, reviewCount]
    ProductDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        status:
          type: string
        shortDescription:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDesc:
          type: ["string", "null"]
        price:
          type: string
        compareAtPrice:
          type: ["string", "null"]
        cost:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        weightUnit:
          type: string
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        isDigital:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        faqs:
          type: array
          items:
            $ref: "#/components/schemas/ProductFAQ"
        brand:
          $ref: "#/components/schemas/BrandSummary"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"
        reviewCount:
          type: integer
        avgRating:
          type: number
      required: [id, name, slug, sku, status, price, weightUnit, isFeatured, isNewArrival, isDigital, createdAt, updatedAt, images, variants, specs, faqs, categories, reviewCount, avgRating]
    ProductImage:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        altText:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required: [id, url, sortOrder]
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        price:
          type: ["string", "null"]
        compareAtPrice:
          type: ["string", "null"]
        inventory:
          type: integer
        isAvailable:
          type: boolean
        options:
          type: object
        imageUrl:
          type: ["string", "null"]
      required: [id, sku, name, inventory, isAvailable]
    ProductSpec:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        value:
          type: string
      required: [id, label, value]
    ProductFAQ:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
      required: [id, question, answer]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductWithRelations"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        status:
          type: string
          enum: [DRAFT, ACTIVE, ARCHIVED]
        shortDescription:
          type: string
        description:
          type: string
        price:
          type: string
        compareAtPrice:
          type: string
        cost:
          type: string
        weight:
          type: string
        weightUnit:
          type: string
        brandId:
          type: string
        distributorId:
          type: string
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        categoryIds:
          type: array
          items:
            type: string
        collectionIds:
          type: array
          items:
            type: string
      required: [name, slug, sku, price]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [DRAFT, ACTIVE, ARCHIVED]
        shortDescription:
          type: string
        description:
          type: string
        price:
          type: string
        compareAtPrice:
          type: string
        cost:
          type: string
        weight:
          type: string
        weightUnit:
          type: string
        brandId:
          type: string
        distributorId:
          type: string
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
...
    CategorySummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
      required: [id, name, slug]
...
    BrandSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
      required: [id, name, slug]
...
    Review:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        authorName:
          type: string
        rating:
          type: integer
        title:
          type: ["string", "null"]
        body:
          type: string
        isVerified:
          type: boolean
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, productId, authorName, rating, body, isVerified, isApproved, createdAt]
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [reviews, total, page, limit, totalPages]
    CreateReviewBody:
      type: object
      properties:
        productId:
          type: string
        authorName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
      required: [productId, authorName, rating, body]
    UpdateReviewBody:
      type: object
      properties:
        isApproved:
          type: boolean
        isVerified:
          type: boolean
...